Git Repository Creation
Two versions of face_perception_behavioral.R
writeSingleXLSX outputs an excel file. This is not the most useful
functionality for the whole study, but this is a nice proof of concept.

Will need to figure out how to call functions from other R scripts to do
the big processing stuff.